Fixed some printf's bugging on linux. Added linux visuals by default for demonstrating

+++ Quantom/Quantom/BoxBuilder.cu
@@ -94,8 +94,8 @@ void BoxBuilder::finishBox(Simulation* simulation) {
 	// Permanent Outputs for energy & trajectory analysis
 	int n_points = simulation->total_particles_upperbound * STEPS_PER_LOGTRANSFER;
 	printf("n points %d\n", n_points);
-	printf("Malloc %.2Lf KB on device for data buffers\n", (sizeof(double) * simulation->total_particles_upperbound * STEPS_PER_LOGTRANSFER + sizeof(Float3) * simulation->total_particles_upperbound * STEPS_PER_LOGTRANSFER) * 1e-3);
-	printf("Malloc %.2Lf MB on host for data buffers\n", (sizeof(double) * simulation->total_particles_upperbound * STEPS_PER_LOGTRANSFER + sizeof(Float3) * simulation->total_particles_upperbound * STEPS_PER_LOGTRANSFER) * 1e-6);
+	printf("Malloc %.2f KB on device for data buffers\n",(float) ((sizeof(double) * simulation->total_particles_upperbound * STEPS_PER_LOGTRANSFER + sizeof(Float3) * simulation->total_particles_upperbound * STEPS_PER_LOGTRANSFER) * 1e-3));
+	printf("Malloc %.2f MB on host for data buffers\n", (float) ((sizeof(double) * simulation->total_particles_upperbound * STEPS_PER_LOGTRANSFER + sizeof(Float3) * simulation->total_particles_upperbound * STEPS_PER_LOGTRANSFER) * 1e-6));
 	cudaMallocManaged(&simulation->box->potE_buffer, sizeof(double) * simulation->total_particles_upperbound * STEPS_PER_LOGTRANSFER);	// Can only log molecules of size 3 for now...
 	simulation->potE_buffer = new double[simulation->total_particles_upperbound * simulation->n_steps];
 
@@ -117,7 +117,7 @@ void BoxBuilder::finishBox(Simulation* simulation) {
 	uint64_t n_traindata_device = N_DATAGAN_VALUES * MAX_COMPOUND_PARTICLES * simulation->n_compounds * STEPS_PER_TRAINDATATRANSFER;
 	long double total_bytes = sizeof(float) * n_loggingdata_device
 		+ sizeof(Float3) * n_traindata_device;
-	printf("Reserving %.2Lf MB device mem for logging + training data\n", (total_bytes) * 1e-6);
+	printf("Reserving %.2f MB device mem for logging + training data\n", (float) ((total_bytes) * 1e-6));
 	cudaMallocManaged(&simulation->box->outdata, sizeof(float) * 10 * STEPS_PER_LOGTRANSFER);	// 10 data streams for 10k steps. 1 step at a time.
 
 	cudaMallocManaged(&simulation->box->data_GAN, sizeof(Float3) * N_DATAGAN_VALUES * MAX_COMPOUND_PARTICLES * simulation->n_compounds * STEPS_PER_TRAINDATATRANSFER);
@@ -125,7 +125,7 @@ void BoxBuilder::finishBox(Simulation* simulation) {
 
 	uint64_t n_loggingdata_host = 10 * simulation->n_steps;
 	uint64_t n_traindata_host = N_DATAGAN_VALUES * MAX_COMPOUND_PARTICLES * simulation->n_compounds * (uint64_t) simulation->n_steps;
-	printf("Reserving %.2Lf GB host mem for logging + training data\n", (sizeof(Float3) * n_traindata_host + sizeof(float) * n_loggingdata_host) * 1e-9);
+	printf("Reserving %.2f GB host mem for logging + training data\n",(float) (sizeof(Float3) * n_traindata_host + sizeof(float) * n_loggingdata_host) * 1e-9);
 	simulation->logging_data = new float[n_loggingdata_host];
 	simulation->traindata_buffer = new Float3[n_traindata_host];
